Successful response from model to view
There is a working response from the model to the view. The view sends a
simple request to the controller which sends it to the model and then
the model sends it back to the controller which sends it to the view.
The view just alerts what it was sent.